Moved classes to data_models/
