+++ summarization/extractor_output_preprocessor.py
@@ -10,34 +10,9 @@ from sentence_transformers import SentenceTransformer
 
 from data_models.contents import ContentType
 from data_models.preprocessed_contents import PreprocessedContents
+from data_models.sentence_with_attributes import SentenceWithAttributes
 from summarization.text_summarization import TextSummarizer
 from summarization.web_entity_detection import ImageDescriptionRetriever
-
-
-class SentenceWithAttributes:
-    ''' Object to represent a summary sentence
-    or title sentence along with attributes such as
-    embedding or font_style
-    '''
-
-    def __init__(
-            self,
-            text,
-            para_index,
-            sentence_index_in_para,
-            sentence_weight,
-            font_style,
-            embedding):
-        self.text = text
-        self.paragraph_index = para_index
-        self.sentence_index_in_para = sentence_index_in_para
-        self.sentence_weight = sentence_weight
-        self.font_style = font_style
-        self.embedding = embedding
-
-    def get_weighted_index(self):
-        return self.paragraph_index \
-            + self.sentence_index_in_para * self.sentence_weight
 
 
 class ExtractorOutputPreprocessor:
